+++ lib/strong_migrations/adapters/mysql_adapter.rb
@@ -44,7 +44,7 @@ module StrongMigrations
       end
 
       def add_column_default_safe?
-        server_version >= Gem::Version.new("8.0.12")
+        true
       end
 
       def change_type_safe?(table, column, type, options, existing_column, existing_type)