+++ test/add_column_test.rb
@@ -2,30 +2,35 @@ require_relative "test_helper"
 
 class AddColumnTest < Minitest::Test
   def test_default
-    with_target_version(postgresql? ? 10 : (mysql? ? "8.0.11" : "10.3.1")) do
-      assert_unsafe AddColumnDefault
+    if mysql? || mariadb?
+      assert_safe AddColumnDefault
+    else
+      with_target_version(10) do
+        assert_unsafe AddColumnDefault
+      end
     end
   end
 
   def test_default_database_specific_versions
+    skip unless postgresql?
     StrongMigrations.target_postgresql_version = "10"
-    StrongMigrations.target_mysql_version = "8.0.11"
-    StrongMigrations.target_mariadb_version = "10.3.1"
     assert_unsafe AddColumnDefault
   ensure
     StrongMigrations.target_postgresql_version = nil
-    StrongMigrations.target_mysql_version = nil
-    StrongMigrations.target_mariadb_version = nil
   end
 
   def test_default_null
-    with_target_version(postgresql? ? 10 : (mysql? ? "8.0.11" : "10.3.1")) do
+    skip unless postgresql?
+
+    with_target_version(10) do
       assert_unsafe AddColumnDefaultNull, "Adding a column with a null default"
     end
   end
 
   def test_default_not_null
-    with_target_version(postgresql? ? 10 : (mysql? ? "8.0.11" : "10.3.1")) do
+    skip unless postgresql?
+
+    with_target_version(10) do
       assert_unsafe AddColumnDefaultNotNull, "Then add the NOT NULL constraint"
     end
   end